Some Minor Fixes
Accumulate in BokehDepthOfField takes a float4 for in/out, but FragBlur expects to provide it half4 and the function itself processes the relevant values as half4. This fixes it.

There's no need for the XR Legacy Input Helpers in the Universal Test Project, so I've removed it from the manifest.